requirements.txt
sentence-transformers
torch
faiss-cpu
tqdm
numpy
pandas
spacy
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl
transformers
accelerate
datasets
shap